[feat][#54] listArea에 path 추가 및 안쓰는 뷰 코드 옮기기
